Fix high CPU usage and multi-monitor bug
* Fixing high CPU usage: When the spotify DBUS is not available, we would attempt
to try connect to it aggressively on the command thread. This caused high CPU
usage. We have introduced a small delay after each attempt to prevent that.

* Fixing multi-monitor support: It appears that when Polybar is used on multi-monitor
setups, it spawned different polybars. Consequently, it created a contention for
the socket address we were using. This commit ensures that different socket ports
will be used.